requirements.txt
# Core dependencies for CodePilot

# Ingestion & file processing
chardet==5.2.0

# Embeddings & vector search
sentence-transformers==2.3.1
faiss-cpu==1.7.4
torch==2.1.2

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Utilities
numpy==1.26.3

# Development & Testing
pytest==7.4.3

